+++ patches/src/main/kotlin/li/auna/patches/instagram/interaction/bio/Fingerprints.kt
@@ -1,0 +1,8 @@
+package li.auna.patches.instagram.interaction.bio
+
+import app.revanced.patcher.fingerprint
+
+internal val selectableBioFingerprint = fingerprint {
+    returns("V")
+   strings("is_bio_visible")
+}