+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.7.0-dev.3](https://github.com/Aunali321/ReVancedExperiments/compare/v1.7.0-dev.2...v1.7.0-dev.3) (2025-01-07)
+
+
+### Features
+
+* **instagram:** Add `Max quality` patch ([4c893c9](https://github.com/Aunali321/ReVancedExperiments/commit/4c893c958c2de3acea036ec092101bfdd1ea68c6))
+
+# [1.7.0-dev.2](https://github.com/Aunali321/ReVancedExperiments/compare/v1.7.0-dev.1...v1.7.0-dev.2) (2025-01-07)
+
+
+### Features
+
+* **instagram:** Add `Developer menu` patch ([29b5f46](https://github.com/Aunali321/ReVancedExperiments/commit/29b5f4682b803bf55a1092ae6634f8ec0289d005))
+
+# [1.7.0-dev.1](https://github.com/Aunali321/ReVancedExperiments/compare/v1.6.0...v1.7.0-dev.1) (2025-01-06)
+
+
+### Features
+
+* **instagram:** Add `SelectableBioPatch` patch ([a20f5ae](https://github.com/Aunali321/ReVancedExperiments/commit/a20f5ae11a1ecaab5deb56e6d46f2efbf2f30a64))
+
 # [1.6.0](https://github.com/Aunali321/ReVancedExperiments/compare/v1.5.0...v1.6.0) (2024-12-11)